moved openvpn/discovery -> services/openvpn/discovery
