+++ services/openvpn/discovery/dto/bandwidth.go
