+++ services/openvpn/discovery/dto/service_definition.go
