+++ services/openvpn/discovery/service_poposal_test.go
@@ -23,8 +23,8 @@ import (
 
 	nats_discovery "github.com/mysteriumnetwork/node/communication/nats/discovery"
 	"github.com/mysteriumnetwork/node/openvpn"
-	dto_openvpn "github.com/mysteriumnetwork/node/openvpn/discovery/dto"
 	dto_discovery "github.com/mysteriumnetwork/node/service_discovery/dto"
+	dto_openvpn "github.com/mysteriumnetwork/node/services/openvpn/discovery/dto"
 	"github.com/stretchr/testify/assert"
 )
